Support for ATMega328PB (#15)
* Add support for ATMega328PB

* Update readme

* Update readme

* Update readme

* Fix bug

* Update readme

* Fix bug

* Cosmetics

* Fix bug

* Fix timer2 scaler

* Fix bug

* Fix timer2 scaler

* Prevent matrix out of bounds

* Cosmetics

* Cosmetics

+++ msrc/formatData.cpp
@@ -54,7 +54,7 @@ uint32_t FormatData::formatLatLon(uint8_t type, float value)
     if (type == TYPE_LON) {
         data |= (uint32_t)1 << 31;
     }
-    data |= abs(round(value * 10000));
+    data |= (uint32_t)abs(round(value * 10000));
     return data;
 }
